requirements.txt
djangorestframework
pyyaml
requests
django-cors-headers


